- added support for MongoDB - Get and Post now save and retrieve data from db - change BigInteger which stores numbers within MultiplicationTable with Int64 due to lack of support from MongoDB - added DbService.cs which takes care of connecting to database - added FibonacciModelMongoDB which is used for storing objects from db and saving them to db - added connection string to confing - TODO-clean up commented code, move database name to proper place from DbService.cs

+++ backapp/FibonacciRestApi/Libs/CustomMath.cs
@@ -7,17 +7,17 @@ namespace CustomMath
 {
     public static class Fibonacci
     {
-        public static List<BigInteger> GetSequence(int? sequenceLength)
+        public static List<Int64> GetSequence(int? sequenceLength)
         {
-            BigInteger firtValue = 0;
-            BigInteger secondValue = 1;
+            Int64 firtValue = 0;
+            Int64 secondValue = 1;
             return GetSequence(firtValue, secondValue, sequenceLength);
         }
 
-        public static List<BigInteger> GetSequence(BigInteger firstValue, BigInteger secondValue, int? sequenceLength)
+        public static List<Int64> GetSequence(Int64 firstValue, Int64 secondValue, int? sequenceLength)
         {
-            List<BigInteger> sequence = new List<BigInteger>();
-            BigInteger tempValue = 0;
+            List<Int64> sequence = new List<Int64>();
+            Int64 tempValue = 0;
 
             if(sequenceLength > 0)
             {
@@ -42,26 +42,26 @@ namespace CustomMath
             return sequence;
         }
 
-        public static List<List<BigInteger>> GetMultiplicationTable(int? sequenceLength)
+        public static List<List<Int64>> GetMultiplicationTable(int? sequenceLength)
         {
-            BigInteger firstValue = 0;
-            BigInteger secondValue = 1;
+            Int64 firstValue = 0;
+            Int64 secondValue = 1;
             return GetMultiplicationTable(firstValue, secondValue, sequenceLength);
         }
 
-        public static List<List<BigInteger>> GetMultiplicationTable(BigInteger firstValue, BigInteger secondValue, int? sequenceLength)
+        public static List<List<Int64>> GetMultiplicationTable(Int64 firstValue, Int64 secondValue, int? sequenceLength)
         {
-            List<List<BigInteger>> multiplicationTable = new List<List<BigInteger>>();
+            List<List<Int64>> multiplicationTable = new List<List<Int64>>();
             if(sequenceLength > 0)
             {
-                List<BigInteger> fibonacciSequence = GetSequence(firstValue, secondValue, sequenceLength);
+                List<Int64> fibonacciSequence = GetSequence(firstValue, secondValue, sequenceLength);
 
                 multiplicationTable.Add(fibonacciSequence);
 
                 for (int i = 0; i < fibonacciSequence.Count; i++)
                 {
-                    List<BigInteger> tableRow = new List<BigInteger>();
-                    BigInteger baseNumber = fibonacciSequence[i];
+                    List<Int64> tableRow = new List<Int64>();
+                    Int64 baseNumber = fibonacciSequence[i];
                     tableRow.Add(baseNumber);
 
                     for (int j = 0; j < fibonacciSequence.Count; j++)
